+++ pyxem/tests/signals/test_pixelated_stem_class.py
@@ -255,384 +255,6 @@ class TestDiffraction2DThresholdAndMask:
             s.threshold_and_mask()
 
 
-class TestDiffraction2DCenterOfMass:
-    def test_center_of_mass_0d(self):
-        x0, y0 = 2, 3
-        array0 = np.zeros(shape=(7, 9))
-        array0[y0, x0] = 1
-        s0 = Diffraction2D(array0)
-        s_com0 = s0.center_of_mass()
-        assert (s_com0.inav[0].data == x0).all()
-        assert (s_com0.inav[1].data == y0).all()
-        assert s_com0.axes_manager.navigation_shape == (2,)
-        assert s_com0.axes_manager.signal_shape == ()
-
-    def test_center_of_mass_1d(self):
-        x0, y0 = 2, 3
-        array0 = np.zeros(shape=(5, 6, 4))
-        array0[:, y0, x0] = 1
-        s0 = Diffraction2D(array0)
-        s_com0 = s0.center_of_mass()
-        assert (s_com0.inav[0].data == x0).all()
-        assert (s_com0.inav[1].data == y0).all()
-        assert s_com0.axes_manager.navigation_shape == (2,)
-        assert s_com0.axes_manager.signal_shape == (5,)
-
-    def test_center_of_mass(self):
-        x0, y0 = 5, 7
-        array0 = np.zeros(shape=(10, 10, 10, 10))
-        array0[:, :, y0, x0] = 1
-        s0 = Diffraction2D(array0)
-        s_com0 = s0.center_of_mass()
-        assert (s_com0.inav[0].data == x0).all()
-        assert (s_com0.inav[1].data == y0).all()
-
-        array1 = np.zeros(shape=(10, 10, 10, 10))
-        x1_array = np.random.randint(0, 10, size=(10, 10))
-        y1_array = np.random.randint(0, 10, size=(10, 10))
-        for i in range(10):
-            for j in range(10):
-                array1[i, j, y1_array[i, j], x1_array[i, j]] = 1
-        s1 = Diffraction2D(array1)
-        s_com1 = s1.center_of_mass()
-        assert (s_com1.data[0] == x1_array).all()
-        assert (s_com1.data[1] == y1_array).all()
-
-    def test_center_of_mass_different_shapes(self):
-        array1 = np.zeros(shape=(5, 10, 15, 8))
-        x1_array = np.random.randint(1, 7, size=(5, 10))
-        y1_array = np.random.randint(1, 14, size=(5, 10))
-        for i in range(5):
-            for j in range(10):
-                array1[i, j, y1_array[i, j], x1_array[i, j]] = 1
-        s1 = Diffraction2D(array1)
-        s_com1 = s1.center_of_mass()
-        assert (s_com1.inav[0].data == x1_array).all()
-        assert (s_com1.inav[1].data == y1_array).all()
-
-    def test_center_of_mass_different_shapes2(self):
-        psX, psY = 11, 9
-        s = mdtd.generate_4d_data(probe_size_x=psX, probe_size_y=psY, ring_x=None)
-        s_com = s.center_of_mass()
-        assert s_com.axes_manager.shape == (2, psX, psY)
-
-    def test_different_shape_no_blur_no_downscale(self):
-        y, x = np.mgrid[75:83:9j, 85:95:11j]
-        s = mdtd.generate_4d_data(
-            probe_size_x=11,
-            probe_size_y=9,
-            ring_x=None,
-            image_size_x=160,
-            image_size_y=140,
-            disk_x=x,
-            disk_y=y,
-            disk_r=40,
-            disk_I=20,
-            blur=False,
-            blur_sigma=1,
-            downscale=False,
-        )
-        s_com = s.center_of_mass()
-        assert (s_com.inav[0].data == x).all()
-        assert (s_com.inav[1].data == y).all()
-
-    def test_different_shape_no_downscale(self):
-        y, x = np.mgrid[75:83:9j, 85:95:11j]
-        s = mdtd.generate_4d_data(
-            probe_size_x=11,
-            probe_size_y=9,
-            ring_x=None,
-            image_size_x=160,
-            image_size_y=140,
-            disk_x=x,
-            disk_y=y,
-            disk_r=40,
-            disk_I=20,
-            blur=True,
-            blur_sigma=1,
-            downscale=False,
-        )
-        s_com = s.center_of_mass()
-        np.testing.assert_allclose(s_com.inav[0].data, x)
-        np.testing.assert_allclose(s_com.inav[1].data, y)
-
-    def test_mask(self):
-        y, x = np.mgrid[75:83:9j, 85:95:11j]
-        s = mdtd.generate_4d_data(
-            probe_size_x=11,
-            probe_size_y=9,
-            ring_x=None,
-            image_size_x=160,
-            image_size_y=140,
-            disk_x=x,
-            disk_y=y,
-            disk_r=40,
-            disk_I=20,
-            blur=False,
-            blur_sigma=1,
-            downscale=False,
-        )
-        s.data[:, :, 15, 10] = 1000000
-        s_com0 = s.center_of_mass()
-        s_com1 = s.center_of_mass(mask=(90, 79, 60))
-        assert not (s_com0.inav[0].data == x).all()
-        assert not (s_com0.inav[1].data == y).all()
-        assert (s_com1.inav[0].data == x).all()
-        assert (s_com1.inav[1].data == y).all()
-
-    def test_mask_2(self):
-        x, y = 60, 50
-        s = mdtd.generate_4d_data(
-            probe_size_x=5,
-            probe_size_y=5,
-            ring_x=None,
-            image_size_x=120,
-            image_size_y=100,
-            disk_x=x,
-            disk_y=y,
-            disk_r=20,
-            disk_I=20,
-            blur=False,
-            downscale=False,
-        )
-        # Add one large value
-        s.data[:, :, 50, 30] = 200000  # Large value to the left of the disk
-
-        # Center of mass should not be in center of the disk, due to the
-        # large value.
-        s_com0 = s.center_of_mass()
-        assert not (s_com0.inav[0].data == x).all()
-        assert (s_com0.inav[1].data == y).all()
-
-        # Here, the large value is masked
-        s_com1 = s.center_of_mass(mask=(60, 50, 25))
-        assert (s_com1.inav[0].data == x).all()
-        assert (s_com1.inav[1].data == y).all()
-
-        # Here, the large value is right inside the edge of the mask
-        s_com3 = s.center_of_mass(mask=(60, 50, 31))
-        assert not (s_com3.inav[0].data == x).all()
-        assert (s_com3.inav[1].data == y).all()
-
-        # Here, the large value is right inside the edge of the mask
-        s_com4 = s.center_of_mass(mask=(59, 50, 30))
-        assert not (s_com4.inav[0].data == x).all()
-        assert (s_com4.inav[1].data == y).all()
-
-        s.data[:, :, 50, 30] = 0
-        s.data[:, :, 80, 60] = 200000  # Large value under the disk
-
-        # The large value is masked
-        s_com5 = s.center_of_mass(mask=(60, 50, 25))
-        assert (s_com5.inav[0].data == x).all()
-        assert (s_com5.inav[1].data == y).all()
-
-        # The large value just not masked
-        s_com6 = s.center_of_mass(mask=(60, 50, 31))
-        assert (s_com6.inav[0].data == x).all()
-        assert not (s_com6.inav[1].data == y).all()
-
-        # The large value just not masked
-        s_com7 = s.center_of_mass(mask=(60, 55, 25))
-        assert (s_com7.inav[0].data == x).all()
-        assert not (s_com7.inav[1].data == y).all()
-
-    def test_threshold(self):
-        x, y = 60, 50
-        s = mdtd.generate_4d_data(
-            probe_size_x=4,
-            probe_size_y=3,
-            ring_x=None,
-            image_size_x=120,
-            image_size_y=100,
-            disk_x=x,
-            disk_y=y,
-            disk_r=20,
-            disk_I=20,
-            blur=False,
-            blur_sigma=1,
-            downscale=False,
-        )
-        s.data[:, :, 0:30, 0:30] = 5
-
-        # The extra values are ignored due to thresholding
-        s_com0 = s.center_of_mass(threshold=2)
-        assert (s_com0.inav[0].data == x).all()
-        assert (s_com0.inav[1].data == y).all()
-
-        # The extra values are not ignored
-        s_com1 = s.center_of_mass(threshold=1)
-        assert not (s_com1.inav[0].data == x).all()
-        assert not (s_com1.inav[1].data == y).all()
-
-        # The extra values are not ignored
-        s_com2 = s.center_of_mass()
-        assert not (s_com2.inav[0].data == x).all()
-        assert not (s_com2.inav[1].data == y).all()
-
-    def test_threshold_and_mask(self):
-        x, y = 60, 50
-        s = mdtd.generate_4d_data(
-            probe_size_x=4,
-            probe_size_y=3,
-            ring_x=None,
-            image_size_x=120,
-            image_size_y=100,
-            disk_x=x,
-            disk_y=y,
-            disk_r=20,
-            disk_I=20,
-            blur=False,
-            blur_sigma=1,
-            downscale=False,
-        )
-        s.data[:, :, 0:30, 0:30] = 5
-        s.data[:, :, 1, -2] = 60
-
-        # The extra values are ignored due to thresholding and mask
-        s_com0 = s.center_of_mass(threshold=3, mask=(60, 50, 50))
-        assert (s_com0.inav[0].data == x).all()
-        assert (s_com0.inav[1].data == y).all()
-
-        # The extra values are not ignored
-        s_com1 = s.center_of_mass(mask=(60, 50, 50))
-        assert not (s_com1.inav[0].data == x).all()
-        assert not (s_com1.inav[1].data == y).all()
-
-        # The extra values are not ignored
-        s_com3 = s.center_of_mass(threshold=3)
-        assert not (s_com3.inav[0].data == x).all()
-        assert not (s_com3.inav[1].data == y).all()
-
-        # The extra values are not ignored
-        s_com4 = s.center_of_mass()
-        assert not (s_com4.inav[0].data == x).all()
-        assert not (s_com4.inav[1].data == y).all()
-
-    def test_1d_signal(self):
-        x = np.arange(45, 45 + 9).reshape((1, 9))
-        y = np.arange(55, 55 + 9).reshape((1, 9))
-        s = mdtd.generate_4d_data(
-            probe_size_x=9,
-            probe_size_y=1,
-            ring_x=None,
-            image_size_x=120,
-            image_size_y=100,
-            disk_x=x,
-            disk_y=y,
-            disk_r=20,
-            disk_I=20,
-            blur=False,
-            blur_sigma=1,
-            downscale=False,
-        )
-        s_com = s.inav[:, 0].center_of_mass()
-        assert (s_com.inav[0].data == x).all()
-        assert (s_com.inav[1].data == y).all()
-
-    def test_0d_signal(self):
-        x, y = 40, 51
-        s = mdtd.generate_4d_data(
-            probe_size_x=1,
-            probe_size_y=1,
-            ring_x=None,
-            image_size_x=120,
-            image_size_y=100,
-            disk_x=x,
-            disk_y=y,
-            disk_r=20,
-            disk_I=20,
-            blur=False,
-            blur_sigma=1,
-            downscale=False,
-        )
-        s_com = s.inav[0, 0].center_of_mass()
-        assert (s_com.inav[0].data == x).all()
-        assert (s_com.inav[1].data == y).all()
-
-    def test_lazy(self):
-        y, x = np.mgrid[75:83:9j, 85:95:11j]
-        s = mdtd.generate_4d_data(
-            probe_size_x=11,
-            probe_size_y=9,
-            ring_x=None,
-            image_size_x=160,
-            image_size_y=140,
-            disk_x=x,
-            disk_y=y,
-            disk_r=40,
-            disk_I=20,
-            blur=True,
-            blur_sigma=1,
-            downscale=False,
-        )
-        s_lazy = LazyDiffraction2D(da.from_array(s.data, chunks=(1, 1, 140, 160)))
-        s_lazy_com = s_lazy.center_of_mass()
-        np.testing.assert_allclose(s_lazy_com.inav[0].data, x)
-        np.testing.assert_allclose(s_lazy_com.inav[1].data, y)
-
-        s_lazy_1d = s_lazy.inav[0]
-        s_lazy_1d_com = s_lazy_1d.center_of_mass()
-        np.testing.assert_allclose(s_lazy_1d_com.inav[0].data, x[:, 0])
-        np.testing.assert_allclose(s_lazy_1d_com.inav[1].data, y[:, 0])
-
-        s_lazy_0d = s_lazy.inav[0, 0]
-        s_lazy_0d_com = s_lazy_0d.center_of_mass()
-        np.testing.assert_allclose(s_lazy_0d_com.inav[0].data, x[0, 0])
-        np.testing.assert_allclose(s_lazy_0d_com.inav[1].data, y[0, 0])
-
-    def test_compare_lazy_and_nonlazy(self):
-        y, x = np.mgrid[75:83:9j, 85:95:11j]
-        s = mdtd.generate_4d_data(
-            probe_size_x=11,
-            probe_size_y=9,
-            ring_x=None,
-            image_size_x=160,
-            image_size_y=140,
-            disk_x=x,
-            disk_y=y,
-            disk_r=40,
-            disk_I=20,
-            blur=True,
-            blur_sigma=1,
-            downscale=False,
-        )
-        s_lazy = LazyDiffraction2D(da.from_array(s.data, chunks=(1, 1, 140, 160)))
-        s_com = s.center_of_mass()
-        s_lazy_com = s_lazy.center_of_mass()
-        np.testing.assert_equal(s_com.data, s_lazy_com.data)
-
-        com_nav_extent = s_com.axes_manager.navigation_extent
-        lazy_com_nav_extent = s_lazy_com.axes_manager.navigation_extent
-        assert com_nav_extent == lazy_com_nav_extent
-
-        com_sig_extent = s_com.axes_manager.signal_extent
-        lazy_com_sig_extent = s_lazy_com.axes_manager.signal_extent
-        assert com_sig_extent == lazy_com_sig_extent
-
-    def test_lazy_result(self):
-        data = da.ones((10, 10, 20, 20), chunks=(10, 10, 10, 10))
-        s_lazy = LazyDiffraction2D(data)
-        s_lazy_com = s_lazy.center_of_mass(lazy_result=True)
-        assert s_lazy_com._lazy
-        assert s_lazy_com.axes_manager.signal_shape == (10, 10)
-
-        s_lazy_1d = s_lazy.inav[0]
-        s_lazy_1d_com = s_lazy_1d.center_of_mass(lazy_result=True)
-        assert s_lazy_1d_com._lazy
-        assert s_lazy_1d_com.axes_manager.signal_shape == (10,)
-
-        s_lazy_0d = s_lazy.inav[0, 0]
-        s_lazy_0d_com = s_lazy_0d.center_of_mass(lazy_result=True)
-        assert s_lazy_0d_com._lazy
-        assert s_lazy_0d_com.axes_manager.signal_shape == ()
-
-    def test_center_of_mass_inplace(self):
-        with pytest.raises(ValueError):
-            d = Diffraction2D(np.zeros((10, 10, 20, 20)))
-            d.center_of_mass(inplace=True)
-
-
 class TestDiffraction2DAngleSector:
     def test_get_angle_sector_mask_simple(self):
         array = np.zeros((10, 10, 10, 10))
@@ -650,12 +272,12 @@ class TestDiffraction2DAngleSector:
         array = np.zeros((5, 4, 10, 20))
         array[:, :, y, x] = 1
         s = Diffraction2D(array)
-        s_com = s.center_of_mass()
+        s_bs = s.get_direct_beam_position(method="center_of_mass")
         s.angular_mask(
             0.0,
             0.5 * np.pi,
-            centre_x_array=s_com.inav[0].data,
-            centre_y_array=s_com.inav[1].data,
+            centre_x_array=s_bs.isig[0].data,
+            centre_y_array=s_bs.isig[1].data,
         )
 
 
@@ -795,19 +417,12 @@ class TestDiffraction2DShiftDiffraction:
             blur=True,
             ring_x=None,
         )
-        s_com = s.center_of_mass()
-        s_com.data[0] -= centre_x
-        s_com.data[1] -= centre_y
+        s_bs = s.get_direct_beam_position(method="center_of_mass")
         s_shift = s.shift_diffraction(
-            shift_x=s_com.inav[0].data, shift_y=s_com.inav[1].data
+            shift_x=-s_bs.isig[0].data, shift_y=-s_bs.isig[1].data
         )
-        s_shift_c = s_shift.center_of_mass()
-        np.testing.assert_allclose(
-            s_shift_c.data[0], np.ones_like(s_shift_c.data[0]) * centre_x
-        )
-        np.testing.assert_allclose(
-            s_shift_c.data[1], np.ones_like(s_shift_c.data[1]) * centre_y
-        )
+        s_shift_c = s_shift.get_direct_beam_position(method="center_of_mass")
+        assert s_shift_c.data == pytest.approx(0.0)
 
     def test_inplace(self):
         s = Diffraction2D(np.zeros((10, 10, 30, 40)))
