+++ pyxem/utils/plotting.py
@@ -18,6 +18,7 @@
 
 """Utils for plotting 2D Diffraction Patterns."""
 
+import math
 import matplotlib.pyplot as plt
 import numpy as np
 from pyxem.utils.polar_transform_utils import (
@@ -26,6 +27,8 @@ from pyxem.utils.polar_transform_utils import (
     image_to_polar,
 )
 from pyxem.utils.diffraction import find_beam_center_blur
+import pyxem.utils._beam_shift_tools as bst
+from mpl_toolkits.axes_grid1.anchored_artists import AnchoredSizeBar
 
 __all__ = [
     "plot_template_over_pattern",
@@ -137,3 +140,116 @@ def plot_template_over_pattern(
         color=marker_color,
     )
     return (ax, im, sp)
+
+
+def plot_beam_shift_color(
+    signal,
+    phase_rotation=0,
+    indicator_rotation=0,
+    only_phase=False,
+    autolim=True,
+    autolim_sigma=4,
+    magnitude_limits=None,
+    scalebar_size=None,
+    ax=None,
+    ax_indicator=None,
+):
+    """Make a matplotlib figure showing beam shift.
+
+    Parameters
+    ----------
+    phase_rotation : float, default 0
+        Changes the phase of the plotted data.
+        Useful for correcting scan rotation.
+    indicator_rotation : float, default 0
+        Changes the color wheel rotation.
+    only_phase : bool, default False
+        If False, will plot both the magnitude and phase.
+        If True, will only plot the phase.
+    autolim : bool, default True
+    autolim_sigma : float, default 4
+    magnitude_limits : tuple of floats, default None
+        Manually sets the value limits for the color signal.
+        For this, autolim needs to be False.
+    scalebar_size : int, optional
+    ax : Matplotlib subplot, optional
+    ax_indicator : Matplotlib subplot, optional
+        If None, generate a new subplot for the indicator.
+        If False, do not include an indicator
+
+    Examples
+    --------
+    >>> s = pxm.data.dummy_data.get_simple_beam_shift_signal()
+    >>> fig = pxm.utils.plotting.plot_beam_shift_color(s)
+    >>> fig.savefig("simple_beam_shift_test_signal.png")
+
+    Only plotting the phase
+
+    >>> fig = pxm.utils.plotting.plot_beam_shift_color(s, only_phase=True)
+    >>> fig.savefig("simple_beam_shift_test_signal.png")
+
+    Matplotlib subplot as input
+
+    >>> import matplotlib.pyplot as plt
+    >>> fig, ax = plt.subplots()
+    >>> ax_indicator = fig.add_subplot(331)
+    >>> fig = pxm.utils.plotting.plot_beam_shift_color(
+    ...     s, scalebar_size=10, ax=ax, ax_indicator=ax_indicator)
+
+    """
+    indicator_rotation = indicator_rotation + 60
+    if ax is None:
+        set_fig = True
+        fig, ax = plt.subplots(1, 1, figsize=(7, 7))
+    else:
+        fig = ax.figure
+        set_fig = False
+    if only_phase:
+        s = signal.get_phase_signal(rotation=phase_rotation)
+    else:
+        s = signal.get_color_signal(
+            rotation=phase_rotation,
+            autolim=autolim,
+            autolim_sigma=autolim_sigma,
+            magnitude_limits=magnitude_limits,
+        )
+    s.change_dtype("uint16")
+    s.change_dtype("float64")
+    extent = s.axes_manager.navigation_extent
+    extent = [extent[0], extent[1], extent[3], extent[2]]
+    ax.imshow(s.data / 65536.0, extent=extent)
+    if ax_indicator is not False:
+        if ax_indicator is None:
+            ax_indicator = fig.add_subplot(331)
+        _make_color_wheel(
+            ax_indicator, rotation=indicator_rotation + phase_rotation
+        )
+    ax.set_axis_off()
+    if scalebar_size is not None:
+        scalebar_label = "{0} {1}".format(scalebar_size, s.axes_manager[0].units)
+        sb = AnchoredSizeBar(ax.transData, scalebar_size, scalebar_label, loc=4)
+        ax.add_artist(sb)
+    if set_fig:
+        fig.subplots_adjust(0, 0, 1, 1)
+    return fig
+
+
+def _make_color_wheel(ax, rotation=None):
+    x, y = np.mgrid[-2.0:2.0:500j, -2.0:2.0:500j]
+    r = (x**2 + y**2) ** 0.5
+    t = np.arctan2(x, y)
+    del x, y
+    if rotation is not None:
+        t += math.radians(rotation)
+        t = (t + np.pi) % (2 * np.pi) - np.pi
+
+    r_masked = np.ma.masked_where((2.0 < r) | (r < 1.0), r)
+    r_masked -= 1.0
+
+    mask = r_masked.mask
+    r_masked.data[r_masked.mask] = r_masked.mean()
+    rgb_array = bst._get_rgb_phase_magnitude_array(t, r_masked.data)
+    rgb_array = np.dstack((rgb_array, np.invert(mask)))
+
+    ax.imshow(rgb_array, interpolation="quadric", origin="lower")
+    ax.set_axis_off()
